feat: add total chart - exsising charts refactor
