+++ src/components/Effort.tsx
@@ -12,8 +12,10 @@ import { COOKIE_TOKEN_KEY } from '../utils/constants';
 import { getCookie, removeCookie } from '../utils/cookie';
 import { getDateLabelFromMonthYear } from '../utils/dates';
 import { Effort as TEffort } from '../utils/types';
-import { Charts } from './Charts';
 import { Month } from './Month';
+import { MonthChart } from './MonthChart';
+import { TotalChart } from './TotalChart';
+import { t } from '../locale/labels';
 
 export const Effort = component$(() => {
 	const appStore = useContext(AppContext);
@@ -106,19 +108,28 @@ export const Effort = component$(() => {
 					);
 				})}
 			</div>
-			{monthYearListSig.value.map((monthYear, key) => {
-				return (
-					<div key={key} class='m-4'>
-						<div class='text-lg font-bold'>
-							{getDateLabelFromMonthYear(monthYear)}
+			<div class='m-4'>
+				<div class='text-lg font-bold'>{t('total')}</div>
+				<TotalChart
+					monthYearList={monthYearListSig}
+					effort={filteredEffortSig}
+				/>
+			</div>
+			<section class='grid grid-cols-2'>
+				{monthYearListSig.value.map((monthYear, key) => {
+					return (
+						<div key={key} class='m-4'>
+							<div class='text-lg font-bold'>
+								{getDateLabelFromMonthYear(monthYear)}
+							</div>
+							<MonthChart
+								monthYear={monthYear}
+								effort={filteredEffortSig}
+							/>
 						</div>
-						<Charts
-							monthYear={monthYear}
-							effort={filteredEffortSig}
-						/>
-					</div>
-				);
-			})}
+					);
+				})}
+			</section>
 		</div>
 	);
 });
