+++ src/locale/it.json
@@ -14,5 +14,6 @@
 	"empty": "Empty",
 	"gravatar": "Gravatar.com",
 	"gravatarAccount": "create a Gravatar account",
-	"noteProfileImage": "We use #linkGravatar for our profile images.<div> To set or change yours, #accountGravatar.</div>"
+	"noteProfileImage": "We use #linkGravatar for our profile images.<div> To set or change yours, #accountGravatar.</div>",
+	"total": "Total"
 }
